pping_mwe: Update README and small fixes
Update the README with among other things a table containing the
number of instructions the verifier processes for various versions of
the program.

Also add INCLUDE-defines for the XDP and tc programs (so one of them
can easily be removed), and add a small filter to the bpftrace
program to handle null-pointers.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>