pping: Add RTT-based sampling
Add an option (-R, --rtt-rate) to adapt the rate sampling based on the
RTT of the flow. The sampling rate will be C * RTT, where C is a
configurable constant (ex 1.0 to get one sample every RTT), and RTT
is either the current minimum (default) or smoothed RTT of the
flow (chosen via the -T or --rtt-type option).

The smoothed RTT (sRTT) is updated for each calculated RTT, and is
calculated in a similar manner to srtt in the kernel's TCP stack. The
sRTT is a moving average of all RTTs, and is calculated according to
the formula:

  srtt = 7/8 * prev_srtt + 1/8 * rtt

To allow the user to pass a non-integer C (ex 0.1 to get 10 RTT
samples for every RTT-period), fixed-point arithmetic have been used
in the eBPF programs (due to lack of support for floats). The maximum
value for C has been limited to 10000 in order for it to be unlikely
that the C * RTT calculation will overflow (with C = 10000, overflow
will only occur if RTT > 28 seconds).

Concerns with this commit:
- If RTT-based rate sampling is used, the update rate of flow RTT
  stats depends on the current value of the RTT. A flow with a high
  RTT will create RTT samples less frequently, and thus the RTT of
  that flow  will also update less frequently. This could potentially
  be undesired/unexpected behavior.
  - A consequence of this is that the sRTT could generally be expected
    to more quickly adept to growing RTTs than shrinking RTTs, as it
    will be updated more frequently if it starts with lower RTT
    values. The minimum RTT does not have this issue as it does not
    grow.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>