pping: Do both timestamping and matching on ingress and egress
Perform both timestamping and matching on both ingress and egress
hooks. This makes it more similar to Kathie's pping, allowing the tool
to capture RTTs in both directions when deployed on just a single
interface.

Like Kathie's pping, by default filter out RTTs for packets going to
the local machine (will only include local processing delays). This
behavior can be disabled by passing the -l/--localfilt-off option.

As packets that are timestamped on ingress and matched on egress will
include the local machines processing delay, add the "match_on_egress"
member to the JSON output that can be used to differentiate between
RTTs that include the local processing delay, and those which don't.

Finally, report the source and destination addresses from the perspective
of the reply packet, rather than the timestamped packet, to be
consistent with Kathie's pping.

Overall, refactor large parts of pping_kern to allow both timestamping
and matching, as well as updating both the flow and reverse flow and
handle flow-events related to them, in one go. Also update README to
reflect changes.

Concerns with this commit:
- Performance may be worse due to the increased complexity of handling
  both directions of the flow. Additionally, if local-filtering is
  used (enabled by default), will also have to perform a fib-lookup.
- For the standard and ppviz output formats, it's not possible to tell
  if the reported RTT includes delays from local processing or not.
- No longer works kernel 5.4 (verifier seems to object against lookup
  of config.use_srtt).

Other notes:
- Verifier seems to have a much easier time verifying the refactored
  code, "only" processing 156k instructions (down from ~850k) despite
  the overall increase in complexity.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>