pping: Add option to aggregate RTTs
Add an option -a or --aggregate to provide an aggregate report of RTT
samples every X seconds. This is currently mutually exclusive with the
normal per-RTT sample reports.

The aggregated stats are never reset, and thus contain all RTTs since
the start of tracing. The next commit will change this to reset the
stats after every report, so that each report only contain the RTTs
since the last report.

The RTTs are aggregated and reported per IP-prefix. Right now the user
space component only sets a single global IP-prefix (::/0) which
catches all IPs. Later commits will add the option for the user to
provide custom IP-prefixes.

In this intital implementation for aggregating RTTs, the minimum and
maximum RTT are tracked and all RTTs are added to a histogram. It uses
a predetermined number of bins of equal width (by default 1000 bins,
each 1 ms wide), see RTT_AGG_NR_BINS and RTT_AGG_BIN_WIDTH in
pping.h. In the future this can be changed to use more sophisticated
histograms that better capture a wide variety of RTTs.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>