pping: Use atomic add for aggregated bin counts
While the packets from the same flow are likely to be processed on the
same CPU, the same cannot be said for all packets from a certain IP
prefix. Therefore, the aggregated stats (which are performed per
IP-prefix) are much more prone to concurrent updates than other logic
which is performed per flow. It therefore makes sense to guard the
aggregated state from concurrent updates. Use __sync_fetch_and_add()
to increment the bin counts for the aggregated RTT stats.

Note that the aggregated minimum and maximum RTT may still suffer from
lost updates in case of concurrency. As there is no per-CPU version of
the LPM map and LPM maps currently don't support spin locks, there is
currently no convenient way to prevent this. However, as the RTTs are
sampled anyways (either due to the sample rate limit, or due the TCP
timestamps not updating for every packet), we are never guaranteed to
capture see the minimum/maximum RTT in the first place.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>