pping: Add switch for aggregating by dst IP instead of src
By default ePPing will aggregate RTTs based on the src IP of the reply
packet. I.e. the RTT A->B->A will be aggregated based on IP of B. In some
scenarios it may be more interesting to aggregate based on the dst IP
of the reply packet (IP of A in above example). Therefore, add a
switch (-d/--agg-by-dst) which makes ePPing aggregate RTTs based on
the dst IP of the reply packet instead of the src IP.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>