pping: Filter traffic based on IP prefixes
When aggregating RTTs per configured IP-prefix, some of the traffic
may not end up going to or comming from one of the IP prefixes, and
any RTTs from this traffic will thus not get aggregated. It is
unnecessary for ePPing to process traffic that it will not include in
its output. Therefore, add filtering so that ePPing only processes
traffic for which the RTT will be aggregated into one of the
IP-prefixes.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>