pping: Add a dummy mode
Add a dummy mode in which the eBPF programs are empty (i.e. they only
pass the packet on to the kernel). Note that the dummy mode will still
attach these empty programs, and the userspace logic will run as
usual. This indirectly allows testing the overhead of the eBPF
programs, as one can test with/without dummy mode to see the
difference

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>