pping: Add JSON format for aggregation
Add support for output the aggregated reports in JSON format. This
format includes the raw histogram bin counts, making it possible to
post-process the aggregated rtt statistics.

The user specified format (-F/--format) now applies to both event and
aggregated output. If the user attempts to use the ppviz format for
the aggregated output (which is not supported) the program will error
out.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>