pping: Make IP-prefix parsing more robust
The parsing of IP-prefixes was quite fragile/very strict on the
format, and would fail if there were any comments or even any empty
lines in the prefix file. Remedy this by only parsing the start of
each line (ignoring anything after the first space or a hashtag) and
ignoring empty lines.

Furthermore, the functions were constructed so that the parsing had to
be done twice (once for each of the two aggregation maps), which was
inefficient. Remedy this by adding each parsed IP-prefix to every
provided map-fd (instead of reparsing the file for each map-fd).

I intend to squash this commit into the one which added the IP-prefix
parsing before merge.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>