pping: WIP - add option to disable LPM filter
Add an option "disable-LPM-filt" which disables the filtering that's
normally performed to ensure ePPing only processes packets for one of
the provided IP-prefixes. While the filtering can save overhead (and
space for map entries) for traffic that doesn't go towards one of the
monitored IP-prefixes, it adds some additional overhead for traffic
that does go towards one of the monitored IP-prefixes. By allowing the
user to disable the filtering, the user can reduce the overhead in
case it knows that all (or most) of the traffic will be going to one
of the monitored IP-prefixes.

TODO: Automatically disable the filtering if the global
IP-prefix (::/0) is added, as then all traffic will always be
monitored. Ideally it would also be disabled if all IP-prefixes
combined together cover the entire possible address space, but
detecting this is somewhat complicated.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>