pping: Minor cleanup of argument parsing
The parse_arguments() function used to have a separate variable for
each float (or rather a double) value it would parse from the user. As
only one argument will be parsed at the time this is redudant and will
require more and more variables as new options are added. Replace all
these variables with a single "user_val", which is used for all
options that parse a float from the user.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>