pping: Add option to aggregate RTTs
Add an option -a or --aggregate to provide an aggregate report of RTT
samples every X seconds. This is currently mutually exclusive with the
normal per-RTT sample reports.

The aggregated stats are never reset, and thus contain all RTTs since
the start of tracing. The next commit will change this to reset the
stats after every report, so that each report only contain the RTTs
since the last report.

The RTTs are aggregated and reported per IP-prefix, where the user can
modify the size of the prefixes used for IPv4 and IPv6 using the
--aggregate-subnet-v4/v6 flags.

In this intital implementation for aggregating RTTs, the minimum and
maximum RTT are tracked and all RTTs are added to a histogram. It uses
a predetermined number of bins of equal width (set to 1000 bins,
each 1 ms wide), see RTT_AGG_NR_BINS and RTT_AGG_BIN_WIDTH in
pping.h. In the future this can be changed to use more sophisticated
histograms that better capture a wide variety of RTTs.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>