pping: Add option to truncate histogram output
For the json format, the entire histogram is included in the
output. However, in many cases the upper range of the histogram may be
empty, and thus it will print a lot of trailing zeros. Add an option
to not print out these trailing zeros, and thus only print out the
bins up to the last non-empty one.

Using this option can potentially save a lot of space when RTTs are
well below the maximum one that would fit in the histogram. Unlike
removing leading empty bins or ranges of empty bins in the middle of
the histogram, it also does not require encoding any additional
information (like how many bins have been removed) to be possible to
process correctly later.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>