pping: Define map sizes
Instead of specifying the map size directly in the map definitions,
add them as defines at the top of the file to make them easier to
change (don't have to find the correct map among the 12 currently
defined in the file). For the aggregation maps it also avoids having
to specify the size 4 times.

While at it, increase the default packet_ts to 131072 (2^17) entries,
as the previous value of 16384 (2^14) which, especially for the
packet_ts map, was fairly underdimensioned. If only half of the
timestamps are echoed back (due to ex. delayed ACK), it would in
theory be enough with just 16k / (500 * 1) = 32 concurrent flows to
fill it up with stale entries (assuming default cleanout interval of
1s). Increasing the size of these maps will increase the corresponding
memory cost from 2^14 * (48 + 4) = 832 KiB and 2^14 * (44 + 144) =
2.94 MiB to 2^17 * (48 + 4) = 6.5 MiB and 2^17 * (44 + 144) = 23.5
MiB, respectively, which should be acceptable. Compared to the
aggregation maps (which could range in sizes from around 0.25 - 32 GB
depending on no. of CPU cores), the memory footprint of the packet_ts
and flow_state maps are still minimal.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>