pping: Add option to write to file
Add an option -w/--write <filename> which writes the output to the
provided file instead of to stdout. While not particularly useful in
and of itself (the same effect can be achieved by just redirecting
stdout to a file), later commits can add options that are not easily
achievable through redirection. An example would be to "rotate" the
output file, so that not all data goes into a singular large file.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>