pping: Add wraparound option for the file rotation
Add an option -W/--wrap-rotation <N> which allows the user to limit
the maximum amount of different files that can be created by the
-G/--rotate-seconds option, similar to tcpdump -W. Once the rotation
has created N unique files, the next time the file is rotated it will
overwrite the first of the created files, and then the second first
and so on. In that way, only the N most recent files are kept.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>