Add quick test of tc RX timestamps
Add a small program to just test out the tc RX timestamps.

Was too lazy to write a proper loader, so use a simple script instead
relying on iproute loader.

Had to update the linux/bpf.h header as the __sk_buff tstamp_type and
hwtstamp are relatively new fields.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>