pping: Add option to configure tc priority
Add an the option -p/--tc-priority which allows the user to configure
which priority the tc BPF programs will be attached with on the
interface. This allows for more deterministicly configuring how ePPing
should interact with other tc BPF programs (run before/after) instead
of just leaving it up to the kernel to assign a priority (which seems
to assign priority based on which order you attach the BPF programs).

If no priority is assigned, the default value is 0 which means to let
the kernel decide (as previously).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>