pping: Add option to write to file
Add an option -w/--write <filename> which writes the output to the
provided file instead of to stdout. Fail if file the provided file
already exists to avoid data loss (if truncating file) or corrupting
data (if appending file, JSON is not concatable).

To ensure it fails at opening a file that already exists, use the "x"
alternative for fopen (requires C11).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>