pping: Reopen output file on SIGHUP
Make the user space process reopen the output file (if it's writing to
file with the -w/--write option) when it recieves a SIGHUP
signal. This makes it possible to for example rotate the output files
with logrotate.

Note that in case the program receives a SIGHUP signal BEFORE the
output file has been renamed, it will fail to reopen the file (as it
will fail opening existing files to avoid corrupting/losing
data). This will cause the program to shutdown.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>