pping: Add additional per-prefix packet counters
When running in aggregated mode (-a/--aggregate), the previous
per-prefix packet and byte counters only included traffic that the RTT
was tracked for, i.e. by default only TCP traffic with TCP
timestamps (which a flowstate could be created for) was counted. This
makes it hard to correlate RTTs with traffic load, as the total
traffic load to/from a given prefix is not known.

Therefore, split up the per-prefix counters into 3 sets of counters:
- One for TCP traffic with timestamps, i.e. the TCP traffic we can
  track RTTs of
- One of TCP traffic without timestamps, i.e TCP traffic we cannot
  track due to relying on TCP timestamps
- One for non-TCP traffic, which when combined with the other counters
  gives the total amount of (IP) traffic going to/from a prefix

Do NOT create NEW prefix entries for traffic which the RTT cannot be
tracked for. This means that if some prefix only sees traffic of a
type that RTTs cannot be captured for, they will use the global /0
backup entries.

To keep the standard output somewhat manageable (it is already quite
wide), only output the total packet and byte counts for the traffic
to/from the prefix. For the JSON format, output all 3 sets of
counters.

As the aggregated stats now contain a fair bit of information in
addition to the RTT distribution, update names of functions and
structs in the code dealing with the aggregated data to refer to it as
"aggregated_stats" rather than "aggregated_rtts".

This commit will likely considerably increase the overhead for traffic
types that RTT isn't tracked for compared to before. Previously, the
eBPF programs would early abort as soon as it discovered that the
packet was of a type which it couldn't track RTT for. Now, all IP
packets will have their IP-address processed and later used to lookup
the relevant prefixes to update the packet counters for. However, the
overhead for packets that the RTT can't be tracked for should still be
considerbly lower than for packets the RTT can be tracked for, and the
overhead for packets the RTT can be tracked for should not increase
much from previously.

TODO: Run performance tests at end of PR to see how large impact the
changes have both on RTT-trackable and non-trackable traffic.

Potential bug/issue. If the program is configured to NOT track RTTs
for TCP traffic (by using the -I/--icmp flag without the -T/--tcp
flag), the program will not parse the TCP header and thus be unable to
detect if it contains TCP timestamps. Therefore, all TCP packets will
then be counted as TCP packets without timestamps, regardless if they
have timestamps or not.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>