Refactor packet_info structure
Instead of duplicating all the protocol_info members in packet_info
and copying over all those members, make packet_info simply contain a
protocol_info member which it let's the protocol parsing functions
directly fill in.

To allow for this however, the direct packet_id members of packet_info
are removed (as the identifiers are now kept directly in the
protocol_info struct). This should overall results in slightly more
overhead, as when a packet timestamp is added or looked up, the BPF
program now has to copy the flow-tuple into a packet_id struct,
instead of simply using a pointer to the prepared packet_id structs in
packet_info.

TODO: Update comments

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>