pping: Add global per-protocol packet counters
Add an option -g/--global-counters which adds periodical output at the
configured interval containing the total packet and byte counts per
IP-protocol seen since the last report. These counters include all the
packets that the eBPF programs process, regardless if they can be
parsed or not (if a packet cannot be parsed to the IP-level, it is
added to a non-IP counter).

As the ppviz format lacks support for this type of output, error out
in case the user tries to use global-counters with the ppviz format.

Some implementation details:
Internally keep one set of counters for every possible value the
protocol/next-header field in IPv4/IPv6 header, i.e. [0,255].
However, only provide output for non-zero counters. That way
output is minimized while still allowing for detecting unexpected
protocols (or even illegal protocol values). Translate the most
well-known protocol numbers and simply report the protocol-number for
less common protocols (see the proto_to_str() function).

Unlike the per-prefix stats, do not reset the global counters. Instead
keep a copy of the previous counts and calculate the difference in
user space to report the difference since the previous report. This
unsycnhronized approach is simplier than synchronized approach
swapping between two instances of the map used by the per-prefix
stats, but may result in small inconsistencies (ex. the packet-count
and byte-count may missmatch in case the counters are fetched when an
eBPF program has updated the packet-counter but not the byte-counter).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>