pping: Add option to get periodic map utilization stats
Add the --map-stats option, which provides additional output at a
configurable interval with stats on the map utilization for the BPF
maps. This is intended to replace the current debug output (which can
be activated by building with 'DEBUG=1 make').

Support the packet_ts and flowstate maps, with the following stats:
- Number of entries currently in the map (calculate based on the
  total number of entries created and deleted for the map)
- Number of entries created
- Number of entries deleted (only include entries automatically
  deleted by the BPF programs from ex. matching a TSval or seeing a
  FIN/RST flag)
- Number of entries expired (entries deleted for being too old,
  deleted + expired gives you total number of entries removed from the
  BPF map)
- Number of clean cycles (times the map have been iterated through to
  expire old entries)
- Number of elements processed by the clean-cycles
- Runtime spent in the clean-cycles

For all stats expect the number of entries currently in the map,
report the difference since the previous report (e.g. report the
number of entries created since the last report, rather than the total
number of entries created since start).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>