pping: Add aggregation maps to map utilization stats
Add collection of map utilization stats (for the --map-stats option)
for the aggregation maps. Only report the utilization of the
aggregation maps if the aggregated mode (-a/--aggregate) is used.

Note that there are two instances of the IPv4 and IPv6 aggregation
maps which are swapped between to avoid concurrent access from user
space and BPF programs. Always report the utilization for both
instances of the maps, regardless which are the active ones. It would
be nice to hide this implementation detail, however, the two different
map instances may differ from each other, making it hard to accurately
report the utilization (for any arbitrary map utilization report
interval which may not line up with the aggregation interval) without
reporting both instances.

To enable the reporting cleaning/expiring stats for the aggergation
maps (which are iterated from userspace rather than from BPF iter
programs), add support for collecting and merging userspace map
utilization stats with the ones collected from BPF-space.

Note that the clean-time for the aggregation maps includes the time it
took to report the entries (not just deleting expired
entries). Furthermore, the userspace code will add the backup keys in
the aggregation maps on startup, which will count as the creation of
one entry. Finally, the BPF programs never delete any entries in the
aggregation maps, and thus the deleted field will always be 0, however
it is still included to be consistent with the reported stats for the
other maps.

A small note on formatting:
In the report_aggregated_rtts() function some of the added lines
exceed 80 characters (86). I have intentionally allowed these lines to
exceed this limit, as splitting the line (using clang-format with the
kernel-style config) in my opinion results in harder to read
code. Compare the current formatting of:

ipv4_util = &maputil_ctx->userspace_stats[PPING_MAP_AGG_V4_1];
ipv6_util = &maputil_ctx->userspace_stats[PPING_MAP_AGG_V6_1];

with the line-split version:

ipv4_util =
	&maputil_ctx
		 ->userspace_stats[PPING_MAP_AGG_V4_1];
ipv6_util =
	&maputil_ctx
		 ->userspace_stats[PPING_MAP_AGG_V6_1];

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>