pping: Add aggregation maps to map utilization stats
Add collection of map utilization stats (for the --map-stats option)
for the aggregation maps. Only report the utilization of the
aggregation maps if the aggregated mode (-a/--aggregate) is used.

Note that there are two instances of the IPv4 and IPv6 aggregation
maps which are swapped between to avoid concurrent access from user
space and BPF programs. Only report map utilization stats for the
currently active instance (i.e. the one utilized by the BPF programs)
and use the same name for both instances of the map.

To enable the reporting cleaning/expiring stats for the aggergation
maps (which are iterated from userspace rather than from BPF iter
programs), add support for collecting and merging userspace map
utilization stats with the ones collected from BPF-space.

Note that the clean-time for the aggregation maps includes the time it
took to report the entries (not just deleting expired
entries). Furthermore, the userspace code will add the backup keys in
the aggregation maps on startup, which will count as the creation of
one entry. Finally, the BPF programs never delete any entries in the
aggregation maps, and thus the deleted field will always be 0.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>