pping: Add failure to create entries to map utilization stats
Add a counter for the number of times an entry could not be created
for a map (as the map was full) to the map utilization stats provided
by --map-stats.

The overall count of entries that could not be created is more useful
than the single instance per second provided by the current "map full"
warnings.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>