Add netstacklat tool
Add the tool netstacklat, which measures latency up to different parts
in the Linux ingress network stack. Base the initial implementation on
a bpftrace script from Jesper Dangaard Brouer, which requiers the
kernel to timestamp ingress packets (i.e. set the tstamp member of the
skb). Hence, the latency recorded by the tool is the difference
between the kernel timestamping point and various later points in the
network stack.

In this initial commit, include the eBPF programs for recording the
networks stack latency at the start of the kernel functions
tcp_v4_do_rcv(), tcp_data_queue(), and udp_queue_rcv_one_skb(). Use a
structure making it easy to extend the tool with additional hook
points in the future. Make the eBPF programs compatible with
Cloudflare's ebpf_exporter, and use the map helpers (maps.bpf.h) from
ebpf_exporter to ensure maps are used in a compatible way.

Open code the histogram maps for different hook points as entirely
separate maps, instead of encoding the different hook points in the
key of a separate map as ebpf_exporter often does. This avoids costly
hashmap lookups, as simple array maps can be used instead of hash
maps.

Also include a minimal user space loader, which loads and attaches the
eBPF programs. Later commits will extend this loader to also report
the recorded latencies stored in the BPF maps.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>