netstacklat: Add netstacklat tool
Add the tool netstacklat, which measures latency up to different parts
in the Linux ingress network stack. Base the initial implementation on
a bpftrace script from Jesper Dangaard Brouer, which requiers the
kernel to timestamp ingress packets (i.e. set the tstamp member of the
skb). Hence, the latency recorded by the tool is the difference
between the kernel timestamping point and various later points in the
network stack.

Track the latency at the following points:
- ip-start: Packet has reached the IP-stack (past tc-layer).
  Hooks fentry:ip_rcv_core, ip6_rcv_core
- tcp-start: Packet has reached the local TCP stack (done with all
  IP-processing)
  Hooks: tcp_v4_rcv, tcp_v6_rcv
- udp-start: Packet has reached the local UDP stack (done with all
  IP-processing)
  Hooks: udp_rcv, udpv6_rcv
- tcp-socket-enqueued: Packet has been enqueued to a TCP socket
  Hooks fexit:tcp_data_queue
- udp-socket-enqueued: Packed has been enqueued to a UDP socket
  Hooks fexit: udp_queue_one_skb, udpv6_queue_rcv_one_skb
- tcp-socket-read: Application has read packet data from a TCP socket
  Hooks fentry:tcp_recv_timestamp
- udp-socket-read: Application has read packet data from a UDP socket
  Hooks fentry:skb_consume_udp

Only include a minimal user space loader, which loads and attaches the
eBPF programs, in this initial commit. Later commits will extend this
loader to also report the recorded latencies stored in the BPF maps
and allow it to modify behavior of the eBPF programs.

Make the eBPF programs compatible with Cloudflare's ebpf-exporter [1],
so that statistics can easily be exported as Prometheus metrics. As
part of this, name the BPF maps used to store the histograms according
to Prometheus naming conventions [2]

However, as the macros [3] for storing values in histograms from
ebpf-exporter relies on expanding multiple levels of macros,
reimplement the needed functionality from these macros (specifically
increment_exp2_histogram_nosync) as our own functions to reduce the
risk of running into subtle and hard to debug issues.

However, keep the signature of the increment_exp2_histogram_nosync()
function identical to the corresponding ebpf-exporter macro, making it
possible to replace the function with the ebpf-exporter macros without
changing the call sites. This makes it easy to switch to using the
ebpf-exporter macros instead if desired (e.g. because the macros have
been updated to include some debugging). Notably, this requires
passing the key struct by value, which is otherwise somewhat
unnecessary for the trivial key used by netstacklat.

Note that the increment_exp2_histogram_nosync() function here only
implements the needed functionality by netstacklat. This means that it
only works for ARRAY maps (it assumes all keys already exist, so does
not zero-initialize them as needed for HASH maps), and only works for
a single key type (struct hist_key).

Instead of placing the functions for storing values in histograms in
our own maps.bpf.h header, place them directly in the
netstacklat.bpf.c source. This ensures that if moved to the
ebpf-exporter repository the program will still use these functions
rather than the ebpf-exporter macros, preventing any unforeseen issues
due to macro expansion.

Open code the histogram maps for different hook points as entirely
separate maps, instead of encoding the different hook points in the
key of a single map as ebpf-exporter often does. This allows using
fast array map lookups instead of costly hash map lookup.

[1] https://github.com/cloudflare/ebpf_exporter
[2] https://prometheus.io/docs/practices/naming
[3] https://github.com/cloudflare/ebpf_exporter/blob/master/examples/maps.bpf.h

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>