Make userspace prog report the recorded latencies
Add functionality to the user space component to periodically fetch
the BPF maps netstacklat records the values in and print them out.

Base the core program loop on the same setup as pping, where a single
epoll instance is used to support multiple different types of
events. So far it only deals with signal handling (for clean shutdown)
plus a timer (for periodical reporting), but the setup can easily be
extended if the program grows more complex in the future.

Use the (somewhat complicated) bpf_map_lookup_batch to fetch the
entire histogram maps in a single system call (instead of performing a
lookup for each bin index individually).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>