Use single multiplexed histogram map
Instead of using a separate BPF map for each hook, use a single BPF
map to store the histograms for all hooks by including the hook-id as
part of the key. This is similar to how most existing examples in
ebpf-exporter does it and makes the metric easier to use in
Prometheus, but will likely have higher overhead due to using a
hashmap instead of an array map.

As this is only an experimental change to investigate the (eBPF side)
performance implications of using a hashmap instead of multiple array
maps, remove affected reporting-code in the user space component
instead of making it compatible with the changed map layout.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>