netstacklat: Enable RX software timestamping on startup
The netstacklat tool depends on the tstamp member in the skbs to be
set by the kernel. So make the user space loader enable software RX
timestamping on all packets when the tool is started.

Note that, as the documentation at
https://docs.kernel.org/networking/timestamping.html points out,
enabling software receive timestamping on a socket enables it for all
packets, as the destination socket is not known early in the network
stack. It is therefore sufficient to enable it on a socket created by
the program itself, even if no packets are received on that socket.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>