WIP - netstacklat: Evaluate pid-filter implementations
Evaluate a few different implementations of a pid filter.

These are the tested ones so far:
- Filter single "constant" pid
- Do a binary search over an array of "constant" pids
- Do an array-map lookup
- Do a hash-map lookup

Overall, the array map lookup approach seems quite promising. The main
drawback with it is that it's not very memory efficent as it needs 4M
entries to support the full range of PID values, but this still only
translates to about 32MB of RAM.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>