netstacklat: Add command line argument parsing
Add support for parsing arguments from the command line. Add the
--help/-h and --report-interval/-r <n-seconds> options. As part of
this, add some minor documentation about the program and the available
options that are displayed by the --help option.

Base the print_usage() function on the usage() function from the
traffic-pacing-edt example, but change its output format to be more
consistent with most GNU utilities, listing the short option (if any)
before the long option. Furthermore, also show if an argument is
expected based on the has_arg value.

Automatically generate the optstring that getopt_long() expects from
the options array. While needlessly complicated for the two options
currently supported, it simplifies adding additional options in the
future.

To allow for mapping the opt value returned by getopt_long() back the
corresponding option struct (for example to access the full name of
the option), add the optval_to_longopt() helper that linearly searches
through the options array. This is more reliable than using the
longindex pointer from getopt_long(), as that pointer is only set when
a long option is passed, and not when the corresponding short option
is passed.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>