netstacklat: Track standing socket queues
Keep track if any standing queues are forming at the socket layer. Use
logic similar to CoDel, where a standing (bad) queue is defined as a
queue where the minimum sojourn time exceeds a target over an
interval. Record the durations of the standing queue (i.e. the duration
during which the sojourn time exceeded the target) in a separate
histogram.

Note that while the logic is inspired by CoDel, it is a bit different.
CoDel will act as soon as it detects a standing queue, while we only
want to record a single standing queue duration until the standing
queue is drained. We must therefore wait until after the standing
queue had drained before we can detect it, and therefore need to track
both when it went above (first_above) and below (last_above) the
target in order to avoid including potential idle periods in the
standing queue duration.

To be able to detect if the socket receive queue is empty, need to
read the sock->sk_receive_queue member. So add a definition of struct
sock to vmlinux_net.h, but only include the currently needed members
to avoid having to add the large number of type dependencies that the
full struct sock has, and rely on CO-RE to fix the correct offsets.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>