netstacklat: Multiplex all histograms in a single map
Change the way that the latency histograms are stored on the BPF
side. Instead of keeping a separate array map for each histogram,
store all histograms in a single hash map, encoding the hook as part
of the key.

This results in higher overhead (as hash lookups are slower than array
lookups), but is much more flexible. This makes it easier to add
additional hook points as no new maps (and related code for mapping
hookpoints to maps) need to be added. Furthermore, in the future it
allows to easily group the results on various aspects by adding
additional members to the key.

The userspace agent will be updated to support the new format in a
later commit.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>