netstacklat: Support multiplexed histogram map in user space
Update the userspace code to be able to fetch all the histogram data
from the single histogram hashmap.

Note that while the hashmaps keeps a separate key for each individual
bucket (to be compatible with ebpf-exporter), the userspace agent will
restructure so that the bucket is essentially removed from the key,
keeping a single full histogram (any buckets not found in the map will
be zero-filled) for the remaining unique key members.

Maintain a sorted sequence of the encountered keys (currently just a
statically allocated array). New keys are first appended at the end of
the sequence and swapped into the right (sorted) position, essentially
an insertion sort. Use binary search to quickly lookup the histogram
for a specific key when updating the histogram. If a large number of
unique histograms need to be maintained, may be more efficient to use
some type of hashmap or tree structure in the future (although that
would require adding some suitable data structure library as a
dependency). However, this implementation should be sufficient for as
long we only need to maintain in the order of thousands of unique
histograms.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>