netstacklat: Add filtering for network interfaces
Add the option -i/--interfaces option to filter for specific network
interfaces. The interfaces can either be provided as interface names
or indices, although if the interfaces are in another namespace than
the netstacklat userspace agent is running in they SHOULD be provided
as indices. The names are resolved in the current namespace, and may
therefore fail or yield incorrect indices if the interface is in
another namespace.

Unlike the previous -p/--pids option, this option applies to all
existing probe points in netstacklat. On the eBPF side, use
skb->skb_iif if the skb is available as context, otherwise use the
sk->sk_rx_dst_ifindex from the socket. Use a similar approach as the
previous PID filter, where an array map is used to hold ifindices that
should be filtered for, allowing a quick lookup. While the
ifindex (unlike the PID) does not seem to have a clear upper limit,
limit it to 16384 (IFINDEX_MAX) to keep the filter map reasonably
small while still supporting the vast majority of scenarios.

Note that internally filtering is applied on the interface
index (ifindex), regardless if the option provided the index or the
name for the interface. If the same ifindex is repeated in multiple
network namespaces, it will include traffic for all of them. A future
commit will add an option to also filter for a specific network
namespaces.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>