netstacklat: Generalize initialization of array filter maps
The filtering for specific pids (--pids/-p) makes use of a BPF array
map where each entry to be included is set to 1 (the rest remain 0 as
all entries by default are zeroed in array maps). Generalize the user
space logic that initializes the entries in this filter map so that it
can be reused by other similar features in the future.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>