netstacklat: Move pid-list to heap
The full array of pid-values that could be parsed from the user was
kept inside the config struct, which is allocated on stack (in the
main function). Change the config struct to only keep a pointer to
this array, and allocate it on the heap instead to avoid keeping this
relatively large data structure on the stack.

While not necessarily a large problem yet, establishing this pattern
reduces the risk of running out of stack as new fields to filter for
are added down the line or the maximum number of values to parse is
increased.

Also rename MAX_FIILTER_PIDS to MAX_PARSED_PIDS to better reflect what
it actually is, and update the comment in parse_arguments() to reflect
that the option is called pids and not filter-pids.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>