netstacklat: Add filtering for network namespace
Add the -n/--network-namespace option, which let's the user specify
which network namespace ID (inode number) should be monitored. Apply
the filtering to all current netstacklat probe points.

Use the value 0 (default) to filter for the network namespace that the
netstacklat application itself is running in. Use value -1 to disable
the filtering, including data from all network namespaces (equivalent
with the behavior before this commit).

Only support filtering for a single network namespace (or all
namespaces if the filtering is disabled). This minimizes runtime
overhead by allowing the ID to filter for to be kept as a constant in
the eBPF program. Supporting multiple values would require an
additional map lookup, and due to the wide range of IDs it would have
to be a hashmap lookup, which would add considerable overhead for a
rather niche use case (monitoring multiple network namespaces).

Note that this option will interact with the -i/--interfaces option,
as the ifindex that the --interface option filters for are relative to
the network namespace set by this option.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>