netstacklat: Add filtering for cgroups
Add the -c/--cgroups option that lets the user specify one or more
cgroups (v2) to filter for. The cgroups can either be provided through
their absolute path (including mount path) or as the cgroup
ID (inode). This filter only applies to the probe points running in
process context, just like the PIDs filter, which is currently only
the socket dequeue probes (tcp-read and udp-read).

To keep it simple (and avoid high run-time overhead), only do an exact
match on the cgroup ID. Do not consider the hierarchical relationship
between cgroups. I.e. if the output is filtered for a parent cgroup,
the children of that cgroup will NOT be included unless the children
cgroups have also been explicitly specified.

To support the wide range of possible cgroup IDs, keep the cgroups to
filter for in a sparse hasmap (where only values to include have
entries) instead of the dense array maps (where all possible values
have keys but those to include have non-zero values) like previous
multi-valued filters. This unfortuantely adds considerable overhead
for doing an additional hash map lookup, but keeping a dense map for
all possible IDs is not feasible.

If running with ebpf-exporter, the cgroup filter map can be populated
and maintained by ebpf-exporter by configuring the cgroup_id_map
setting in the YAML config. Update the YAML config to show an example
of how it can be set (in this case including all cgroups under the
system.slice).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>