netstacklat: Exclude TCP reads for HOL blocked segments
The 'tcp-socket-read' currently reports the latency for the skb
containing the last TCP segment read from the socket. However, this
segment might have been head of line (HOL) blocked by a previous
segment missing. In this case, netstacklat's reported latency will
include HOL blocking periods that is dependent on external
factors (such as network packet loss, and network latency impacts
retransmission time). As netstacklat is primarily intended to identify
issues within the local host (in the network stack or receiving
applications), by default filter out any socket reads were the last
read SKB might have experienced HOL-blocking.

Add the new -y/--include-tcp-hol-delay option to retain the old
behavior of reporting latency for all reads, including those that are
HOL-blocked. This may be useful in some scenarios when you still want
to be aware of latency issues caused by HOL-blocking, even though it
is caused by external components. For example, in a data center
context were you have full control over the network, it may still be
relevant to monitor HOL-based caused by the network.

To exclude HOL-blocked reads, track whenever a TCP segment arrives
out-of-order (ooo), i.e. ahead of the expected next sequence. These
segments are added to a separate ooo-queue (by tcp_data_queue_ofo) and
later merged back into the socket receive queue (by tcp_ofo_queue)
once the missing segment gap has been filled. The ooo-segments are
therefore the only segments that may experience HOL blocking (as non
ooo-segments are immediately added to the socket receive
queue). Specifically, track the right edge (the maximum) of the ooo
sequence range. If the final byte of any socket read is lower than
right edge of the ooo sequence range, exclude it as it may potentially
be HOL blocked. If the last read byte is ahead of the ooo-range, the
last byte must have been read from a segment that arrived in-order and
therefore hasn't experienced HOL-blocking, so it's safe to include the
latency measurement. Furthermore, also invalidate the ooo-range when
the last read byte is ahead the ooo-range to prevent an old ooo-range
value sticking around until the sequence number wraps around.

The ideal place to track changes to the ooo-range would be the
tcp_data_queue_ofo() function that adds the skbs to the
ooo-queue. However, the tcp_data_queue_ofo() function may be
untraceable on some systems (due to inlining). Therefore, trace the
end of the tcp_data_queue() function instead (seems to be more
commonly available, even if in theory it can also be inlined), which
is the sole call site of tcp_data_queue_ofo(). If the ooo-queue is
non-empty an skb may be been added to it, so update the ooo-range by
using the end-sequence from the cached tail of the ooo queue (kept in
tcp_sock.ooo_last_skb).

Protect the ooo-range state with a spinlock to prevent concurrency
issues between updates in tcp_data_queue() and reading and
invalidation of the ooo-range in tcp_recv_timestamp(). While
tcp_data_queue() already runs in a context where it is protected by
the socket lock, tcp_recv_timestamp() runs outside of the socket lock,
necessitating our own lock.

There are some scenarios that may cause this ooo-filtering to fail.
- If the program is started in the middle of an ongoing flow, there
  may be ooo segments that arrived before we could record them in the
  the tracked ooo-range. This may cause the latency for some initial
  HOL-blocked reads to be reported.

- If multiple reads are done to the socket concurrently, we may not
  correctly track the last read byte. The kernel does not keep a lock
  on the TCP socket at the time our hooked function
  tcp_recv_timestamp() runs. If two reads are done in parallel, it's
  therefore possible that for both reads we will check the last read
  byte (tcp_sock.copied_seq) after the second read has updated it. We
  may then incorrectly conclude that the first read was ahead of the
  ooo-range when it was not, and record its latency when we should
  have excluded it. In practice I belive this issue should be quite
  rare, as most applications will probably not attempt to perform
  multiple concurrent reads to a single connected TCP socket in
  parallel (as then you cannot know which part of the payload the
  parallel reads will return).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>